fix(gui): CodeQL and Windows-ctest findings from the first full PR run
- cb8_editor: widen the w*h*4 allocation sizes to size_t before the
  multiplication (cpp/integer-multiplication-cast-to-long) — both the
  frame-decode and export paths.
- fonts: drop the WINDIR environment lookup for a fixed probe list
  (cpp/path-injection); non-standard Windows installs fall through to
  the embedded default font like any other miss.
- dialog-queue test: ASCII test name — catch_discover_tests round-trips
  names through ctest and the em dash mangled under the Windows console
  codepage, making Catch2 match zero tests.

The remaining path-injection alert (--smoke LIB arguments) is dismissed
as a false positive: opening user-specified files is the tool's purpose,
same trust model as the fx CLI.

+++ gui/src/editors/cb8_editor.cpp
@@ -33,7 +33,7 @@ static GpuTexture DecodeToTexture(uint32_t frameIdx) {
 
     int w = (int)s_info.width;
     int h = (int)s_info.height;
-    std::vector<uint8_t> rgba(w * h * 4);
+    std::vector<uint8_t> rgba((size_t)w * h * 4);
     for (int i = 0; i < w * h; i++) {
         uint8_t v = indices[i];   // greyscale: index maps to its own value as brightness
         rgba[i*4+0] = v;
@@ -101,7 +101,7 @@ void DrawCb8Editor(App& app) {
             [&app, data = ed.data, info = s_info](std::string dir) {
                 if (dir.empty()) return;
                 int w = (int)info.width, h = (int)info.height;
-                std::vector<uint8_t> rgba(w * h * 4);
+                std::vector<uint8_t> rgba((size_t)w * h * 4);
 
                 // Sequential export with a fresh decoder over the snapshot
                 fx::Cb8Decoder* expDec = fx::cb8_open(data.data(), data.size());
